Buildfix on gdc: add -fall-instantiations
This is to work around an issue with gdc (bug?) where
all!() is not found at link time (if compiling .so)
or at runtime (if compiling .a). This change adds the
-fall-instantiations flag to the library's build,
to the upper level project if this is compiled as a
submodule and to projects linking to this throug pkgconf.